Separate tests related to dead link filtering
