+++ openverse_api/test/dead_link_filter_test.py
@@ -1,6 +1,8 @@
+from test.constants import API_URL
 from unittest.mock import MagicMock, patch
 
 import pytest
+import requests
 
 
 def _patch_redis():
@@ -65,3 +67,71 @@ def test_dead_link_filtering(client):
     res_1_ids = set(result["id"] for result in data_with_dead_links["results"])
     res_2_ids = set(result["id"] for result in data_without_dead_links["results"])
     assert bool(res_1_ids - res_2_ids)
+
+
+@pytest.fixture
+def search_factory(client):
+    """
+    Allows passing url parameters along with a search request.
+    """
+
+    def _parameterized_search(**kwargs):
+        response = requests.get(f"{API_URL}/v1/images", params=kwargs, verify=False)
+        assert response.status_code == 200
+        parsed = response.json()
+        return parsed
+
+    return _parameterized_search
+
+
+@pytest.fixture
+def search_without_dead_links(search_factory):
+    """
+    Here we pass filter_dead = True.
+    """
+
+    def _search_without_dead_links(**kwargs):
+        return search_factory(filter_dead=True, **kwargs)
+
+    return _search_without_dead_links
+
+
+@pytest.mark.django_db
+def test_page_size_removing_dead_links(search_without_dead_links):
+    """
+    We have about 500 dead links in the sample data and should have around
+    8 dead links in the first 100 results on a query composed of a single
+    wildcard operator.
+
+    Test whether the number of results returned is equal to the requested
+    page_size of 100.
+    """
+    data = search_without_dead_links(q="*", page_size=100)
+    assert len(data["results"]) == 100
+
+
+@pytest.mark.django_db
+def test_page_consistency_removing_dead_links(search_without_dead_links):
+    """
+    Test the results returned in consecutive pages are never repeated when
+    filtering out dead links.
+    """
+    total_pages = 30
+    page_size = 5
+
+    page_results = []
+    for page in range(1, total_pages + 1):
+        page_data = search_without_dead_links(q="*", page_size=page_size, page=page)
+        page_results += page_data["results"]
+
+    def no_duplicates(xs):
+        s = set()
+        for x in xs:
+            if x in s:
+                return False
+            s.add(x)
+        return True
+
+    ids = list(map(lambda x: x["id"], page_results))
+    # No results should be repeated so we should have no duplicate ids
+    assert no_duplicates(ids)
